+++ apps/alpha-test-app/src/views/HomeView.vue
@@ -14,14 +14,19 @@ import {
   OnyxLoadingIndicator,
   OnyxPageLayout,
   OnyxRadioButtonGroup,
+  OnyxSelect,
   OnyxSkeleton,
   OnyxSwitch,
   OnyxTooltip,
   type SelectionOption,
 } from "sit-onyx";
-import { computed, ref } from "vue";
+import { computed, ref, capitalize } from "vue";
+import { useI18n } from "vue-i18n";
 import { useRouter } from "vue-router";
+import { DENSITY } from "sit-onyx";
+import LanguageSelection from "../components/LanguageSelection.vue";
 
+const { locale } = useI18n();
 const router = useRouter();
 
 const COMPONENTS = [
@@ -32,12 +37,13 @@ const COMPONENTS = [
   "OnyxIconButton",
   "OnyxInput",
   "OnyxLink",
+  "OnyxListbox",
   "OnyxLoadingIndicator",
   "OnyxRadioButtonGroup",
+  "OnyxSelect",
   "OnyxSkeleton",
   "OnyxSwitch",
   "OnyxTooltip",
-  "OnyxListbox",
 ] as const;
 
 /* Config data to regulate which components will be shown */
@@ -50,6 +56,16 @@ const activeConfig = ref(configOptions.map((option) => option.id));
 const show = computed(() => {
   return (componentName: (typeof COMPONENTS)[number]) => activeConfig.value.includes(componentName);
 });
+
+const densityOptions = DENSITY.map((id) => ({
+  id,
+  label: capitalize(id),
+})) satisfies SelectionOption<string>[];
+
+const activeDensityOption = ref({ ...densityOptions[1] });
+
+const useSkeleton = ref(false);
+const skeletonNumber = computed(() => (useSkeleton.value ? 3 : undefined));
 
 /* Demo data for the components we show */
 const dummyOptions: SelectionOption[] = ["A", "B", "C"].map((id) => ({
@@ -80,6 +96,9 @@ const listboxOptions = [
   "Raspberry",
   "Strawberry",
 ].map<ListboxOption>((option) => ({ id: option.toLowerCase(), label: option }));
+
+const multiSelectState = ref(["Apple", "Banana", "Mango", "Kiwi", "Orange", "Papaya"]);
+const singleSelectState = ref("Apple");
 </script>
 
 <template>
@@ -96,6 +115,15 @@ const listboxOptions = [
     <OnyxPageLayout>
       <template #sidebar>
         <div class="sidebar">
+          <OnyxRadioButtonGroup
+            v-model="activeDensityOption"
+            headline="Density"
+            :options="densityOptions"
+          />
+          <LanguageSelection v-model="locale" />
+
+          <OnyxSwitch v-model="useSkeleton" label="All as Skeleton" />
+
           <OnyxCheckboxGroup
             v-model="activeConfig"
             headline="Examples to show"
@@ -105,21 +133,24 @@ const listboxOptions = [
         </div>
       </template>
 
-      <div class="page">
+      <div class="page" :class="[`onyx-density-${activeDensityOption.id}`]">
         <OnyxHeadline is="h1">Component usages</OnyxHeadline>
 
         <p>Each onyx component should be used at least once in this page.</p>
 
         <div class="page__examples">
-          <OnyxButton v-if="show('OnyxButton')" label="Button" />
+          <OnyxButton v-if="show('OnyxButton')" label="Button" :skeleton="useSkeleton" />
 
           <template v-if="show('OnyxCheckboxGroup')">
             <OnyxCheckboxGroup
               v-model="checkboxState"
               headline="Checkbox Group"
               :options="dummyOptions"
+              :skeleton="skeletonNumber"
             />
-            OnyxCheckboxGroup state: {{ checkboxState }}
+            <div v-if="!useSkeleton" class="onyx-text--small state-info">
+              OnyxCheckboxGroup state: {{ checkboxState }}
+            </div>
           </template>
 
           <OnyxHeadline is="h1" v-if="show('OnyxHeadline')">Headline</OnyxHeadline>
@@ -128,9 +159,16 @@ const listboxOptions = [
 
           <OnyxIconButton v-if="show('OnyxIconButton')" label="Happy Emoji" :icon="emojiHappy2" />
 
-          <OnyxInput v-if="show('OnyxInput')" label="Input" />
+          <OnyxInput v-if="show('OnyxInput')" label="Input" :skeleton="useSkeleton" />
 
-          <OnyxLink v-if="show('OnyxLink')" href="#">Link</OnyxLink>
+          <OnyxLink v-if="show('OnyxLink')" href="#" :skeleton="useSkeleton">Link</OnyxLink>
+
+          <OnyxListbox
+            v-if="show('OnyxListbox')"
+            v-model="listboxState"
+            label="Example listbox"
+            :options="listboxOptions"
+          />
 
           <OnyxLoadingIndicator v-if="show('OnyxLoadingIndicator')" />
 
@@ -139,8 +177,33 @@ const listboxOptions = [
               v-model="radioState"
               headline="Radio Button Group"
               :options="dummyOptions"
+              :skeleton="skeletonNumber"
             />
-            OnyxRadioButtonGroup state: {{ radioState ?? "–" }}
+            <div v-if="!useSkeleton" class="onyx-text--small state-info">
+              OnyxRadioButtonGroup state: {{ radioState ?? "–" }}
+            </div>
+          </template>
+
+          <template v-if="show('OnyxSelect')">
+            <OnyxSelect
+              v-model="singleSelectState"
+              label="Single Select"
+              placeholder="Select your fruits"
+              :skeleton="useSkeleton"
+            />
+            <div v-if="!useSkeleton" class="onyx-text--small state-info">
+              OnyxSelect single state: {{ singleSelectState ?? "–" }}
+            </div>
+            <OnyxSelect
+              v-model="multiSelectState"
+              label="Multiple Select"
+              placeholder="Select your fruits"
+              multiple
+              :skeleton="useSkeleton"
+            />
+            <div v-if="!useSkeleton" class="onyx-text--small state-info">
+              OnyxSelect multiple state: {{ multiSelectState ?? "–" }}
+            </div>
           </template>
 
           <OnyxSkeleton v-if="show('OnyxSkeleton')" class="skeleton" />
@@ -149,15 +212,14 @@ const listboxOptions = [
             v-if="show('OnyxSwitch')"
             v-model="switchState"
             :label="'Switch is ' + (switchState ? 'on' : 'off')"
+            :skeleton="useSkeleton"
           />
 
           <OnyxTooltip v-if="show('OnyxTooltip')" text="Example tooltip text">
             Hover me to show tooltip
           </OnyxTooltip>
 
-          <OnyxListbox v-model="listboxState" label="Example listbox" :options="listboxOptions" />
-
-          <!-- Add new components here. -->
+          <!-- Add new components alphabetically. -->
         </div>
       </div>
     </OnyxPageLayout>
@@ -172,6 +234,9 @@ const listboxOptions = [
   border-bottom: var(--onyx-1px-in-rem) solid var(--onyx-color-base-neutral-300);
 }
 .sidebar {
+  display: flex;
+  flex-direction: column;
+  gap: var(--onyx-spacing-md);
   padding: var(--onyx-spacing-md);
   border-right: var(--onyx-1px-in-rem) solid var(--onyx-color-base-neutral-300);
   height: calc(100% - var(--onyx-spacing-xl));
@@ -189,5 +254,8 @@ const listboxOptions = [
 .skeleton {
   height: 2rem;
   width: 8rem;
+}
+.state-info {
+  color: var(--onyx-color-text-icons-neutral-soft);
 }
 </style>