+++ packages/chartjs-plugin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sit-onyx/chartjs-plugin
 
+## 1.0.0-alpha.15
+
+### Patch Changes
+
+- Updated dependencies [b31b7c5]
+  - sit-onyx@1.0.0-alpha.60
+
+## 1.0.0-alpha.14
+
+### Patch Changes
+
+- Updated dependencies [fe0f615]
+  - sit-onyx@1.0.0-alpha.59
+
 ## 1.0.0-alpha.13
 
 ### Patch Changes