+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,17 @@
 # sit-onyx
 
+## 1.0.0-alpha.60
+
+### Minor Changes
+
+- b31b7c5: feat(OnyxSelect): implement skeleton, readonly, density and translation
+
+## 1.0.0-alpha.59
+
+### Major Changes
+
+- fe0f615: refactor(OnyxTooltip): remove property `label` and `hideLabel` in favor of `message`
+
 ## 1.0.0-alpha.58
 
 ### Minor Changes